adjested directory stucture
